Partial work on merge.  PBDirect runs and all tests pass, but improved support for large classes is missing.
