+++ shared/src/main/scala/pbdirect/package.scala
@@ -21,7 +21,7 @@ package object pbdirect {
   }
   implicit class PBParserOps(private val bytes: Array[Byte]) extends AnyVal {
     def pbTo[A](implicit parser: PBParser[A]): A = {
-      parser.readSingleFieldAndBuild(NEL.one(1), CodedInputStream.newInstance(bytes), size=Some(bytes.length))
+      parser.readSingleFieldAndBuild(CodedInputStream.newInstance(bytes), size=Some(bytes.length))
     }
   }
 }
